stores state in dashboard for isOpen
