+++ src/components/SideBar/index.js
@@ -8,26 +8,7 @@ import Dropdown from "../lib/Dropdown"
 import Flyout from "../lib/Dropdown/Flyout"
 import { history } from "../../store"
 
-const Navigation = ({ user }) => {
-  const [width, setWidth] = useState(window.innerWidth)
-  const [isOpen, setIsOpen] = useState(false)
-  useEffect(() => {
-    const handleResize = () => {
-      return setWidth(window.innerWidth)
-    }
-    window.addEventListener("resize", handleResize)
-    if (width > 768) setIsOpen(true)
-    return () => {
-      window.removeEventListener("resize", handleResize)
-    }
-  }, [width])
-//   const triggerLogout = () => {
-//     localStorage.removeItem("auth")
-//     history.push("/sign-in")
-//   }
-//   const triggerDropdown = () => {
-//     setTimeout(() => setIsOpen(false), 400)
-//   }
+const SideBar = ({ user , isOpen, setIsOpen, width}) => {
 if (width > 768) {
   return <section className="sidebar">
       
@@ -39,4 +20,4 @@ if (width > 768) {
     
   )
 }
-export default Navigation
+export default SideBar
